week 3/rag_app/requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyPDF2
sentence-transformers
pinecone-client
weaviate-client
